Update && remove legacy code

+++ Assets/ScriptableObjects/Weapons/GunData.cs
@@ -7,7 +7,7 @@ public class GunData : ScriptableObject
     public float damage;
     public float radius;
     public float range;
-    public float delay;
+    public int delay;
     public float fireRatePrimary;
     public float fireRateSecondary;
     public float impactForce;
